Update main to dev

+++ mlkaps/sample_collection/subprocess_harness.py
@@ -1,8 +1,8 @@
 """
-    Copyright (C) 2020-2024 Intel Corporation
-    Copyright (C) 2022-2024 University of Versailles Saint-Quentin-en-Yvelines
-    Copyright (C) 2024-  MLKAPS contributors
-    SPDX-License-Identifier: BSD-3-Clause
+Copyright (C) 2020-2024 Intel Corporation
+Copyright (C) 2022-2024 University of Versailles Saint-Quentin-en-Yvelines
+Copyright (C) 2024-  MLKAPS contributors
+SPDX-License-Identifier: BSD-3-Clause
 """
 
 from dataclasses import dataclass
@@ -14,6 +14,8 @@ import signal
 import logging
 import pathlib
 from collections import namedtuple
+from typing import Dict
+import sys
 
 
 @dataclass
@@ -30,14 +32,16 @@ class ProcessCleanupHandler:
     """
     Helper class to cleanly kill all subprocess on failure or user exit
 
-    This class wraps around subprocess.Popen and ensures that the subprocess and its childs are killed if an exception if received.
+    This class wraps around subprocess.Popen and ensures that the subprocess
+    and its childs are killed if an exception if received.
     This includes SystemExit and KeyboardInterrupt, so no zombie process is created if the user kills MLKAPS.
     """
 
     def __init__(self, timeout: float = None):
         """Initiliaze the cleanup handler
 
-        :param timeout: A timeout in seconds for the process to finish. The process will be killed if the timeout expries, defaults to None
+        :param timeout: A timeout in seconds for the process to finish.
+        The process will be killed if the timeout expries, defaults to None
         :type timeout: float, optional
         """
         self.timeout = timeout
@@ -50,9 +54,11 @@ class ProcessCleanupHandler:
         :return: The stdout and stderr of the process
         :rtype: tuple[str, str]
         """
-
-        with suppress(ProcessLookupError):
-            os.killpg(os.getpgid(process.pid), signal.SIGTERM)
+        if os.name == "nt":
+            subprocess.call(["taskkill", "/F", "/T", "/PID", str(process.pid)])
+        else:
+            with suppress(ProcessLookupError):
+                os.killpg(os.getpgid(process.pid), signal.SIGTERM)
         return process.communicate()
 
     def run(self, *args, **kwargs) -> ProcessResult:
@@ -62,24 +68,37 @@ class ProcessCleanupHandler:
         :rtype: tuple[subprocess.Popen, str, str]
         """
 
+        # On Windows, .py file are not executable.
+        # Add "python" as the command. sys.executable is the path to the python interpreter
+
+        if os.name == "nt":
+            assert isinstance(args, tuple)
+            assert len(args) == 1
+            assert isinstance(args[0], list)
+            alist = args[0]
+            wpath = pathlib.WindowsPath(alist[0])
+            if not pathlib.Path(wpath).exists():
+                raise (FileNotFoundError(f"File {wpath} not found"))
+
+            if wpath.suffix == ".py":
+                args = ([sys.executable] + alist,)
+
         timed_out = False
         process = None
         try:
             process = subprocess.Popen(*args, **kwargs)
             stdout, stderror = process.communicate(timeout=self.timeout)
         except subprocess.TimeoutExpired:
-            logging.warning(f"Process timedout, cleaning up...)")
+            logging.warning("Process timedout, cleaning up...)")
 
             stdout, stderror = self._cleanup_kill(process)
             timed_out = True
-        except:
+        except BaseException:  # We should capture all exceptions here to ensure cleanup
             # Can occur if an exception is raised  during Popen
             if process is None:
                 raise
 
-            logging.warning(
-                f"Exception raised during subprocess execution, cleaning up..."
-            )
+            logging.warning("Exception raised during subprocess execution, cleaning up...")
 
             self._cleanup_kill(process)
             raise
@@ -99,11 +118,13 @@ class MonoSubprocessHarness:
     def __init__(
         self,
         objectives: list[str],
+        objectives_bounds: Dict[str, float],
         executable_path: pathlib.Path,
         arguments_order: list[str],
         timeout: float | None = None,
     ):
         self.objectives = objectives
+        self.objectives_bounds = objectives_bounds
         self.executable_path = executable_path
         self.timeout = timeout
         self.arguments_order = arguments_order
@@ -144,16 +165,27 @@ class MonoSubprocessHarness:
 
         arguments = [str(sample[k]) for k in self.arguments_order]
         result = self._run_process(arguments)
-
+        # ----- Set to arbitrary value
+        # UB = {obj: bound for obj, bound in zip(self.objectives, self.bounds)}
+        objectives = None
         if result.timed_out:
             error = f"Process timed out (max {self.timeout} seconds)\n"
+            # ---- save objective to UB, make sense if objective is a time,
+            # would require more thinking for true integration....
+            objectives = {o: self.objectives_bounds[o] for o in self.objectives}  # UB[o]
         elif result.exitcode != 0:
             error = f"Process exited with code {result.exitcode}\n"
+            # ---- save objective to UB, make sense if objective is a time,
+            # would require more thinking for true integration....
+            objectives = {o: self.objectives_bounds[o] for o in self.objectives}  # UB[o]
         else:
             objectives, error = self._parse_output(result.stdout)
 
         if error is not None:
-            objectives = {o: float("nan") for o in self.objectives}
+            # ---- save objective to UB, make sense if objective is a time,
+            # would require more thinking for true integration....
+            objectives = {o: self.objectives_bounds[o] for o in self.objectives}
+            # float("nan") for o in self.objectives}
             msg = textwrap.indent(f"Arguments: {result.arguments}", "\t| ")
             msg += textwrap.indent(f"\nPID: {result.pid}", "\t| ")
             msg += textwrap.indent(f"\nStdout:\n{result.stdout}", "\t> ")
